pyproject.toml
[build-system]
requires = ["setuptools>=70", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bfm-ai-agent"
version = "0.1.0"
description = "BFM AI Agent demo for operational finance monitoring and follow-up automation."
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastapi>=0.115.12",
    "jinja2>=3.1.6",
    "langchain>=0.3.23",
    "langchain-openai>=0.3.12",
    "langfuse>=3.2.4",
    "langgraph>=0.3.31",
    "openpyxl>=3.1.5",
    "pydantic-settings>=2.8.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.40",
    "uvicorn[standard]>=0.34.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.3.5",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-p no:cacheprovider"
